Revert "adrienne/fix: added routing for wallets and wallets build workflow (#10478)"
This reverts commit c355f41a8c8e862499d15c8c8c4ec7001046c223.
