+++ .github/workflows/test.yml
@@ -27,12 +27,11 @@ jobs:
       run: npx tsc --project packages/analytics/tsconfig.json -noEmit
     - name: Check TypeScript for @deriv/stores
       run: npx tsc --project packages/stores/tsconfig.json -noEmit
-    - name: Check TypeScript for @deriv/wallets
-      run: npx tsc --project packages/wallets/tsconfig.json -noEmit
-    - name: Check ESLint for @deriv/wallets
-      run: npx eslint --fix --ignore-path packages/wallets/.eslintignore --config packages/wallets/.eslintrc.js packages/wallets
-    - name: Check Stylelint for @deriv/wallets
-      run: npx stylelint packages/wallets/**/*.scss
+    - name: Check TypeScript and linting for @deriv/wallets
+      run: |-
+        npx tsc --project packages/wallets/tsconfig.json -noEmit
+        npx eslint --fix --config packages/wallets/.eslintrc.js packages/wallets
+        npx stylelint packages/wallets/**/*.scss
     - name: Check tests for @deriv/hooks
       run: bash ./scripts/check-tests.sh packages/hooks/src
     - name: Check tests for @deriv/utils
