+++ packages/appstore/webpack.config.js
@@ -131,10 +131,29 @@ module.exports = function (env) {
                                 resources: [
                                     // eslint-disable-next-line global-require, import/no-dynamic-require
                                     ...require('@deriv/shared/src/styles/index.js'),
+                                    // eslint-disable-next-line global-require, import/no-dynamic-require
+                                    ...require('@deriv/wallets/src/styles/index.js'),
                                 ],
                             },
                         },
                     ],
+                },
+                {
+                    test: /\.svg$/,
+                    issuer: /\/packages\/wallets\/.*(\/)?.*.scss/,
+                    exclude: /node_modules/,
+                    include: /public\//,
+                    type: 'asset/resource',
+                    generator: {
+                        filename: 'appstore/wallets/public/[name].[contenthash][ext]',
+                    },
+                },
+                {
+                    test: /\.svg$/,
+                    issuer: /\/packages\/wallets\/.*(\/)?.*.tsx/,
+                    exclude: /node_modules/,
+                    include: /public\//,
+                    use: svg_loaders,
                 },
                 {
                     test: /\.svg$/,
@@ -147,7 +166,7 @@ module.exports = function (env) {
                 },
                 {
                     test: /\.svg$/,
-                    exclude: /node_modules|public\//,
+                    exclude: [/node_modules|public\//],
                     use: svg_loaders,
                 },
             ],
