+++ packages/core/package.json
@@ -91,7 +91,6 @@
         "@deriv/analytics": "^1.0.0",
         "@deriv/api": "^1.0.0",
         "@deriv/appstore": "^0.0.4",
-        "@deriv/wallets": "^1.0.0",
         "@deriv/bot-web-ui": "^1.0.0",
         "@deriv/cashier": "^1.0.0",
         "@deriv/cfd": "^1.0.0",
