+++ packages/core/src/App/Constants/routes-config.js
@@ -57,13 +57,6 @@ const AppStore = React.lazy(() =>
     })
 );
 
-const Wallets = React.lazy(() =>
-    moduleLoader(() => {
-        // eslint-disable-next-line import/no-unresolved
-        return import(/* webpackChunkName: "wallets" */ '@deriv/wallets');
-    })
-);
-
 const P2P = React.lazy(() =>
     moduleLoader(() => {
         // eslint-disable-next-line import/no-unresolved
@@ -242,12 +235,6 @@ const getModules = () => {
             component: AppStore,
             is_authenticated: true,
             getTitle: () => localize("Trader's Hub"),
-        },
-        {
-            path: routes.wallets,
-            component: Wallets,
-            is_authenticated: true,
-            getTitle: () => localize('Wallets'),
         },
         {
             path: routes.onboarding,
