--- packages/wallets/.eslintignore
@@ -1,2 +1,0 @@
-dist/*.js
-webpack.config.js
