--- packages/wallets/declarations.d.ts
@@ -1,4 +1,0 @@
-declare module '*.svg' {
-    const content: React.ComponentType<React.SVGAttributes<SVGElement>>;
-    export default content;
-}
