+++ packages/wallets/package.json
@@ -2,15 +2,7 @@
     "name": "@deriv/wallets",
     "private": true,
     "version": "1.0.0",
-    "main": "dist/wallets/js/index.js",
-    "engines": {
-        "node": "18.x"
-    },
-    "scripts": {
-        "build": "rimraf dist && NODE_OPTIONS='-r ts-node/register' webpack --progress --config \"./webpack.config.js\"",
-        "serve": "rimraf dist && concurrently \"cross-env BUILD_MODE='serve' NODE_OPTIONS='-r ts-node/register' webpack --progress --watch --config ./webpack.config.js\" \"tsc -w --noEmit --preserveWatchOutput\"",
-        "start": "rimraf dist && npm run test && npm run serve"
-    },
+    "main": "src/index.tsx",
     "dependencies": {
         "@deriv/api": "^1.0.0",
         "@deriv/utils": "^1.0.0",
@@ -20,8 +12,8 @@
         "formik": "^2.1.4",
         "moment": "^2.29.2",
         "qrcode.react": "^1.0.0",
-        "react": "^17.0.2",
         "react-dom": "^17.0.2",
+        "react": "^17.0.2",
         "react-router-dom": "^5.2.0",
         "usehooks-ts": "^2.7.0"
     },
@@ -35,9 +27,6 @@
         "eslint-plugin-simple-import-sort": "^10.0.0",
         "eslint-plugin-sort-destructure-keys": "^1.5.0",
         "eslint-plugin-typescript-sort-keys": "^2.3.0",
-        "typescript": "^4.6.3",
-        "webpack": "^5.81.0",
-        "webpack-cli": "^4.7.2",
-        "webpack-bundle-analyzer": "^4.3.0"
+        "typescript": "^4.6.3"
     }
 }
