+++ packages/wallets/src/components/WalletGradientBackground/WalletGradientBackground.scss
@@ -1140,7 +1140,7 @@
                         top: 0;
                         width: 100%;
                         height: 100%;
-                        background-image: url('../../public/images/wallet-demo-bg-light.svg');
+                        background-image: url('../../public//images/wallet-demo-bg-light.svg');
                         background-repeat: repeat;
                         background-size: 70px;
                         mix-blend-mode: overlay;
