+++ packages/wallets/src/components/WalletListCardActions/WalletListCardActions.tsx
@@ -62,7 +62,7 @@ const WalletListCardActions: React.FC<TProps> = ({ isActive, isDemo, loginid }) 
                                 className='wallets-mobile-actions-content-icon'
                                 key={button.name}
                                 onClick={() => {
-                                    history.push(`/wallets/cashier/${button.name}`);
+                                    history.push(`/appstore/traders-hub/cashier/${button.name}`);
                                 }}
                             >
                                 {button.icon}
@@ -82,7 +82,7 @@ const WalletListCardActions: React.FC<TProps> = ({ isActive, isDemo, loginid }) 
                     key={button.name}
                     onClick={async () => {
                         await switchAccount(loginid);
-                        history.push(`/wallets/cashier/${button.name}`);
+                        history.push(`/appstore/traders-hub/cashier/${button.name}`);
                     }}
                 >
                     {button.icon}
