+++ packages/wallets/src/components/WalletsAccordion/WalletsAccordion.scss
@@ -28,7 +28,7 @@
                 width: 100%;
                 height: 100%;
                 border-radius: 16px;
-                background-image: url('../../public/images/wallet-demo-desktop-bg.svg');
+                background-image: url('../../public//images/wallet-demo-desktop-bg.svg');
                 background-repeat: repeat;
                 background-size: 60px;
                 opacity: 0.24;
