+++ packages/wallets/src/features/cashier/components/WalletCashierContent/WalletCashierContent.tsx
@@ -6,15 +6,15 @@ import { WalletTransfer } from '../../flows/WalletTransfer';
 
 const WalletCashierContent = () => {
     const history = useHistory();
-    const isTransfer = useRouteMatch('/wallets/cashier/transfer');
-    const isDeposit = useRouteMatch('/wallets/cashier/deposit');
-    const isTransactions = useRouteMatch('/wallets/cashier/transactions');
-    const isWithdraw = useRouteMatch('/wallets/cashier/withdraw');
+    const isTransfer = useRouteMatch('/appstore/traders-hub/cashier/transfer');
+    const isDeposit = useRouteMatch('/appstore/traders-hub/cashier/deposit');
+    const isTransactions = useRouteMatch('/appstore/traders-hub/cashier/transactions');
+    const isWithdraw = useRouteMatch('/appstore/traders-hub/cashier/withdraw');
 
     useEffect(() => {
         // redirect to deposit page if no other page is matched
         if (!isTransfer && !isDeposit && !isTransactions && !isWithdraw) {
-            history.push('/wallets/cashier/deposit');
+            history.push('/appstore/traders-hub/cashier/deposit');
         }
     }, [isTransfer, isDeposit, isTransactions, isWithdraw, history]);
