+++ packages/wallets/src/features/cashier/screens/WalletTransactionsScreens/WalletTransactionsNoDataState/WalletTransactionsNoDataState.tsx
@@ -16,13 +16,13 @@ const WalletTransactionsNoDataState = () => {
             <div className='wallets-transactions-no-data-state__buttons'>
                 <button
                     className='wallets-transactions-no-data-state__buttons__transfer'
-                    onClick={() => history.push('/wallets/cashier/transfer')}
+                    onClick={() => history.push('/appstore/traders-hub/cashier/transfer')}
                 >
                     Transfer funds
                 </button>
                 <button
                     className='wallets-transactions-no-data-state__buttons__deposit'
-                    onClick={() => history.push('/wallets/cashier/deposit')}
+                    onClick={() => history.push('/appstore/traders-hub/cashier/deposit')}
                 >
                     Deposit funds
                 </button>
