--- packages/wallets/src/utils/loader/index.ts
@@ -1,2 +1,0 @@
-export * from './lazy-load';
-export * from './module-loader';
