--- packages/wallets/src/utils/loader/module-loader.tsx
@@ -1,16 +1,0 @@
-export const moduleLoader = (lazyComponent: () => Promise<unknown>, attempts = 3, interval = 1500) => {
-    return new Promise((resolve, reject) => {
-        lazyComponent()
-            .then(resolve)
-            .catch((error: unknown) => {
-                // let us retry after 1500 ms
-                setTimeout(() => {
-                    if (attempts === 1) {
-                        reject(error);
-                        return;
-                    }
-                    moduleLoader(lazyComponent, attempts - 1, interval).then(resolve, reject);
-                }, interval);
-            });
-    });
-};
