+++ packages/wallets/tsconfig.json
@@ -1,12 +1,5 @@
 {
     "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "outDir": "./dist",
-        "baseUrl": "./",
-        "paths": {
-            "@deriv/*": ["../*/src"],
-        }
-    },
-    "include": ["src", "./declarations.d.ts", "../../utils.d.ts"],
+    "include": ["src"],
     "exclude": ["../shared/**/*"]
 }
